requirements.txt
google-generativeai>=0.5.0
pydantic>=2.8
python-dotenv>=1.0
pillow>=10.4
opencv-python>=4.10
pytesseract>=0.3.10
pdf2image>=1.17
pypdf>=5.0
reportlab>=4.2
streamlit>=1.37
numpy>=2.0
rich>=13.7
tqdm>=4.66
python-bidi>=0.4.2
